backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
requests==2.28.2
django-money==3.0.0
Pillow==9.3.0
krakenex~=2.1.0
pykrakenapi~=0.3.1